part4 start, organized phonebook backend into modules
